Fix being able to bypass the policy from another namespace
This would happen as the namespace is ANDed together with the
`notPrincipals`, thus if `notPrincipals` is `true`(principal does not
match) and `namespaces` is `false`, it would evaluate the policy to
`false`, thus it would not deny it. As a SA can only be used from one
namespace, then the `notPrincipals` check is sufficient.